+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Updated translations: `bn-BD`, `te-IN`, `hi-IN`, `ta-IN`, `mr-IN`, `bn-IN`, `ti-ET`, `am-ET`
 - Show next button in the registration screens
 - Implement village type ahead in patient entry/edit screens
+- Show overdue count in home screen tabs
 
 ## On Demo
 ### Internal